requirements.txt
requests
pyrogram
python-telegram-bot
python-dotenv
